backend/requirements.txt
ofxparse
pandas
python-dotenv
langchain_community
langchain-ollama
crewai
crewai-tools
websocket
fastapi
uvicorn[standard]
sqlalchemy
numpy
pydantic
openpyxl
jinja2
bcrypt
pyjwt
pymysql
email-validator
python-multipart